refactor: remove old parser
