+++ src/parser/parse.ts
@@ -2,6 +2,8 @@ import { Validation } from '../report'
 import { checkCharType } from './char'
 import {
   CharType,
+  isLetterType,
+  isPunctuationType,
   Mark,
   MutableMark,
   MarkMap,
@@ -11,20 +13,18 @@ import {
   MutableGroupToken,
   MutableToken,
   Token,
-  GroupToken,
-  isLettersType,
-  isPunctuationType
+  GroupToken
 } from './types'
 import {
+  handleLetter,
+  handlePunctuation,
   appendValue,
-  addHyperContent,
+  addRawContent,
   addHyperToken,
   finalizeLastToken,
   getConnectingSpaceLength,
   getHyperMarkMap,
   getPreviousToken,
-  handleContent,
-  handlePunctuation,
   initNewStatus,
   isShorthand,
   handleErrors
@@ -64,10 +64,11 @@ export type MutableParseResult = {
  * Parse a string into several tokens.
  * - half-width content x {1,n} (English words)
  * - full-width content x {1,n} (Chinese sentenses without punctuations in between)
- * - half-width punctuation
- * - width-width punctuation
- * - punctuation pair as special marks: brackets
- * - punctuation pair as a group: quotes
+ * - half-width punctuation -> halfwidth pause or stop punctuation mark
+ * - width-width punctuation -> fullwidth pause or stop punctuation mark
+ * - punctuation pair as special marks: brackets -> bracket
+ * - punctuation pair as a group: quotes -> quotation or book title mark
+ * - -> halfwidth/fullwidth other punctuation mark
  * Besides them there are some special tokens
  * - content-hyper from hyperMarks as input
  * For spaces they would be included as one or multiple successive spaces in
@@ -98,11 +99,11 @@ export const parse = (str: string, hyperMarks: Mark[] = []): ParseResult => {
       // check the next token
       // - if the mark type is raw
       //   - append next token
-      // - else
-      //   - start mark: append token
-      //   - end mark: append token, append mark
+      // - else (the mark type is hyper)
+      //   - start: append token
+      //   - end hyper mark: append token, append mark
       if (hyperMark.type === MarkType.RAW) {
-        addHyperContent(
+        addRawContent(
           status,
           i,
           str.substring(hyperMark.startIndex, hyperMark.endIndex)
@@ -155,12 +156,12 @@ export const parse = (str: string, hyperMarks: Mark[] = []): ParseResult => {
       appendValue(status, char)
     } else if (isPunctuationType(type)) {
       handlePunctuation(i, char, type, status)
-    } else if (isLettersType(type)) {
-      handleContent(i, char, type, status)
+    } else if (isLetterType(type)) {
+      handleLetter(i, char, type, status)
     } else if (type === CharType.EMPTY) {
       // Nothing
     } else {
-      handleContent(i, char, CharType.LETTERS_HALF, status)
+      handleLetter(i, char, CharType.WESTERN_LETTER, status)
     }
   }
   finalizeLastToken(status, str.length)
@@ -180,7 +181,7 @@ const toMutableToken = (token: Token): MutableToken => {
   if (Array.isArray(token)) {
     const mutableToken: MutableGroupToken = token as MutableGroupToken
     mutableToken.modifiedType = token.type
-    mutableToken.modifiedValue = token.content
+    mutableToken.modifiedValue = token.value
     mutableToken.modifiedSpaceAfter = token.spaceAfter
     mutableToken.modifiedStartValue = token.startValue
     mutableToken.modifiedEndValue = token.endValue
@@ -191,7 +192,7 @@ const toMutableToken = (token: Token): MutableToken => {
   } else {
     const mutableToken: MutableSingleToken = token as MutableSingleToken
     mutableToken.modifiedType = token.type
-    mutableToken.modifiedValue = token.content
+    mutableToken.modifiedValue = token.value
     mutableToken.modifiedSpaceAfter = token.spaceAfter
     mutableToken.validations = []
     return mutableToken
