+++ src/parser/util.ts
@@ -11,18 +11,21 @@ import {
   SHORTHAND_CHARS,
   SHORTHAND_PAIR_SET,
   BRACKET_CHAR_SET,
-  QUOTE_CHAR_SET,
+  QUOTATION_CHAR_SET,
   Mark,
   MarkMap,
   MarkSideType,
   MarkType,
-  LettersType,
+  LetterType,
   SingleToken,
   HyperTokenType,
   GroupToken,
   GroupTokenType,
   Token,
-  PunctuationType
+  PunctuationType,
+  SinglePunctuationType,
+  isBracketType,
+  isQuotationType
 } from './types'
 import { ParseStatus } from './parse'
 
@@ -34,6 +37,7 @@ export const handlePunctuation = (
 ): void => {
   // end the last unfinished token
   finalizeLastToken(status, i)
+
   // check the current token type
   // - start of a mark: start an unfinished mark
   // - end of a mark: end the current unfinished mark
@@ -41,49 +45,55 @@ export const handlePunctuation = (
   // - left quote: start a new unfinished group
   // - right quote: end the current unfinished group
   // - other punctuation: add and end the current token
-  if (BRACKET_CHAR_SET.left.indexOf(char) >= 0) {
-    // push (save) the current unfinished mark if have
-    initNewMark(status, i, char)
-    // generate a new token and mark it as a mark punctuation by left
-    // and finish the token
-    addBracketToken(status, i, char, MarkSideType.LEFT)
-  } else if (BRACKET_CHAR_SET.right.indexOf(char) >= 0) {
-    if (!status.lastMark || !status.lastMark.startValue) {
-      addUnmatchedToken(status, i, char)
-      addError(status, i, BRACKET_NOT_OPEN)
-    } else {
-      // generate token as a punctuation
-      addBracketToken(status, i, char, MarkSideType.RIGHT)
-      // end the last unfinished mark
-      // and pop the previous one if exists
-      finalizeCurrentMark(status, i, char)
+  if (isBracketType(type)) {
+    if (BRACKET_CHAR_SET.left.indexOf(char) >= 0) {
+      // push (save) the current unfinished mark if have
+      initNewMark(status, i, char)
+      // generate a new token and mark it as a mark punctuation by left
+      // and finish the token
+      addBracketToken(status, i, char, MarkSideType.LEFT)
+    } else if (BRACKET_CHAR_SET.right.indexOf(char) >= 0) {
+      if (!status.lastMark || !status.lastMark.startValue) {
+        addUnmatchedToken(status, i, char)
+        addError(status, i, BRACKET_NOT_OPEN)
+      } else {
+        // generate token as a punctuation
+        addBracketToken(status, i, char, MarkSideType.RIGHT)
+        // end the last unfinished mark
+        // and pop the previous one if exists
+        finalizeCurrentMark(status, i, char)
+      }
     }
-  } else if (QUOTE_CHAR_SET.neutral.indexOf(char) >= 0) {
-    // - end the last unfinished group
-    // - start a new group
-    if (status.lastGroup && char === status.lastGroup.startValue) {
-      finalizeCurrentGroup(status, i, char)
-    } else {
-      initNewGroup(status, i, char)
-    }
-  } else if (QUOTE_CHAR_SET.left.indexOf(char) >= 0) {
-    initNewGroup(status, i, char)
-  } else if (QUOTE_CHAR_SET.right.indexOf(char) >= 0) {
-    if (!status.lastGroup || !status.lastGroup.startValue) {
-      addUnmatchedToken(status, i, char)
-      addError(status, i, QUOTE_NOT_OPEN)
-    } else {
-      finalizeCurrentGroup(status, i, char)
-    }
-  } else {
-    addNormalPunctuationToken(status, i, char, type)
+    return
   }
+  if (isQuotationType(type)) {
+    if (QUOTATION_CHAR_SET.neutral.indexOf(char) >= 0) {
+      // - end the last unfinished group
+      // - start a new group
+      if (status.lastGroup && char === status.lastGroup.startValue) {
+        finalizeCurrentGroup(status, i, char)
+      } else {
+        initNewGroup(status, i, char)
+      }
+    } else if (QUOTATION_CHAR_SET.left.indexOf(char) >= 0) {
+      initNewGroup(status, i, char)
+    } else if (QUOTATION_CHAR_SET.right.indexOf(char) >= 0) {
+      if (!status.lastGroup || !status.lastGroup.startValue) {
+        addUnmatchedToken(status, i, char)
+        addError(status, i, QUOTE_NOT_OPEN)
+      } else {
+        finalizeCurrentGroup(status, i, char)
+      }
+    }
+    return
+  }
+  addSinglePunctuationToken(status, i, char, type)
 }
 
-export const handleContent = (
+export const handleLetter = (
   i: number,
   char: string,
-  type: LettersType,
+  type: LetterType,
   status: ParseStatus
 ): void => {
   // check if type changed and last token unfinished
@@ -156,10 +166,11 @@ export const finalizeCurrentToken = (
 const markTypeToTokenType = (type: MarkType): HyperTokenType => {
   switch (type) {
     case MarkType.HYPER:
-      return HyperTokenType.HYPER_WRAPPER
+      return HyperTokenType.HYPER_MARK
     case MarkType.BRACKETS:
-      return HyperTokenType.HYPER_WRAPPER_BRACKET
+      return HyperTokenType.BRACKET_MARK
     case MarkType.RAW:
+      // technically never since MarkType.RAW should go to addRawContent()
       return HyperTokenType.INDETERMINATED
   }
 }
@@ -168,14 +179,14 @@ export const addHyperToken = (
   status: ParseStatus,
   index: number,
   mark: Mark,
-  content: string,
+  value: string,
   markSide: MarkSideType
 ) => {
   const token: SingleToken = {
     type: markTypeToTokenType(mark.type),
     index,
-    length: content.length,
-    content: content,
+    length: value.length,
+    value: value,
     spaceAfter: '', // to be finalized
     mark: mark,
     markSide
@@ -183,16 +194,16 @@ export const addHyperToken = (
   finalizeCurrentToken(status, token)
 }
 
-export const addHyperContent = (
+export const addRawContent = (
   status: ParseStatus,
   index: number,
-  content: string
+  value: string
 ) => {
   const token: SingleToken = {
-    type: getHyperContentType(content),
+    type: getHyperContentType(value),
     index,
-    length: content.length,
-    content: content,
+    length: value.length,
+    value: value,
     spaceAfter: '' // to be finalized
   }
   finalizeCurrentToken(status, token)
@@ -228,10 +239,10 @@ export const addBracketToken = (
   markSide: MarkSideType
 ) => {
   const token: SingleToken = {
-    type: HyperTokenType.HYPER_WRAPPER_BRACKET,
+    type: HyperTokenType.BRACKET_MARK,
     index,
     length: 1,
-    content: char,
+    value: char,
     spaceAfter: '', // to be finalized
     mark: status.lastMark,
     markSide
@@ -258,17 +269,17 @@ export const finalizeCurrentMark = (
 
 // normal punctuation
 
-const addNormalPunctuationToken = (
+const addSinglePunctuationToken = (
   status: ParseStatus,
   index: number,
   char: string,
-  type: PunctuationType
+  type: SinglePunctuationType
 ) => {
   const token: SingleToken = {
     type,
     index,
     length: 1,
-    content: char,
+    value: char,
     spaceAfter: '' // to be finalized
   }
   finalizeCurrentToken(status, token)
@@ -283,7 +294,7 @@ const addUnmatchedToken = (
     type: HyperTokenType.UNMATCHED,
     index: i,
     length: 1,
-    content: char,
+    value: char,
     spaceAfter: ''
   }
   finalizeCurrentToken(status, token)
@@ -322,7 +333,7 @@ export const finalizeCurrentGroup = (
   char: string
 ) => {
   if (status.lastGroup) {
-    // index, length, content
+    // index, length, value
     status.lastGroup.endIndex = index
     status.lastGroup.endValue = char
   }
@@ -333,26 +344,26 @@ export const finalizeCurrentGroup = (
   }
 }
 
-// content
+// general content
 
 export const initNewContent = (
   status: ParseStatus,
   index: number,
   char: string,
-  type: LettersType
+  type: LetterType
 ) => {
   status.lastToken = {
     type,
     index,
     length: 1, // to be finalized
-    content: char, // to be finalized
+    value: char, // to be finalized
     spaceAfter: '' // to be finalized
   }
 }
 
 export const appendValue = (status: ParseStatus, char: string) => {
   if (status.lastToken) {
-    status.lastToken.content += char
+    status.lastToken.value += char
     status.lastToken.length++
   }
 }
@@ -410,12 +421,12 @@ export const isShorthand = (
   if (SHORTHAND_CHARS.indexOf(char) < 0) {
     return false
   }
-  if (!status.lastToken || status.lastToken.type !== CharType.LETTERS_HALF) {
+  if (!status.lastToken || status.lastToken.type !== CharType.WESTERN_LETTER) {
     return false
   }
   const nextChar = str[index + 1]
   const nextType = checkCharType(nextChar)
-  if (nextType === CharType.LETTERS_HALF || nextType === CharType.SPACE) {
+  if (nextType === CharType.WESTERN_LETTER || nextType === CharType.SPACE) {
     if (!status.lastGroup) {
       return true
     }
@@ -433,14 +444,14 @@ export const getHyperContentType = (content: string): HyperTokenType => {
   }
   if (content.match(/^<code.*>.*<\/code.*>$/)) {
     // Usually it's <code>...</code>.
-    return HyperTokenType.HYPER_CONTENT_CODE
+    return HyperTokenType.CODE_CONTENT
   }
   if (content.match(/^<.+>$/)) {
     // Usually it's other HTML tags.
     return HyperTokenType.HYPER_CONTENT
   }
   // Usually it's `...`.
-  return HyperTokenType.HYPER_CONTENT_CODE
+  return HyperTokenType.CODE_CONTENT
 }
 
 // error handling
